+++ app/components/UI/TransactionReview/TransactionReviewCard/index.tsx
@@ -1,95 +1,23 @@
-import React, { useState, useCallback } from 'react';
-import { TouchableOpacity, View, StyleSheet, Linking } from 'react-native';
+import React, { useState, useCallback, useEffect } from 'react';
+import { TouchableOpacity, View, Linking } from 'react-native';
 import Summary from '../../../Base/Summary';
 import Text from '../../../Base/Text';
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
-import { isMainnetByChainId, isTestNet } from '../../../../util/networks';
-import { connect } from 'react-redux';
-import PropTypes from 'prop-types';
+import { isMainnetByChainId } from '../../../../util/networks';
 import InfoModal from '../../Swaps/components/InfoModal';
 import FadeAnimationView from '../../FadeAnimationView';
-import SkeletonPlaceholder from 'react-native-skeleton-placeholder';
 import { strings } from '../../../../../locales/i18n';
 import TimeEstimateInfoModal from '../../TimeEstimateInfoModal';
 import useModalHandler from '../../../Base/hooks/useModalHandler';
 import AppConstants from '../../../../core/AppConstants';
 import Device from '../../../../util/device';
-import { useTheme } from '../../../../util/theme';
+import { useAppThemeFromContext, mockTheme } from '../../../../util/theme';
+import { useGasTransaction } from '../../../../core/GasPolling/GasPolling';
+import createStyles from './styles';
+import { TransactionReviewCardProps } from './types';
+import SkeletonComponent from './skeletonComponent';
 
-const createStyles = (colors) =>
-  StyleSheet.create({
-    overview: (noMargin) => ({
-      marginHorizontal: noMargin ? 0 : 24,
-      paddingTop: 10,
-      paddingBottom: 10,
-    }),
-    valuesContainer: {
-      flex: 1,
-      flexDirection: 'row',
-      justifyContent: 'flex-end',
-    },
-    gasInfoContainer: {
-      paddingLeft: 2,
-    },
-    gasInfoIcon: (hasOrigin) => ({
-      color: hasOrigin ? colors.secondary.default : colors.icon.muted,
-    }),
-    amountContainer: {
-      flex: 1,
-      paddingRight: 10,
-    },
-    gasRowContainer: {
-      flexDirection: 'row',
-      flex: 1,
-      alignItems: 'center',
-      marginBottom: 2,
-    },
-    gasBottomRowContainer: {
-      marginTop: 4,
-    },
-    hitSlop: {
-      top: 10,
-      left: 10,
-      bottom: 10,
-      right: 10,
-    },
-    redInfo: {
-      color: colors.error.default,
-    },
-    timeEstimateContainer: {
-      alignItems: 'center',
-      flexDirection: 'row',
-    },
-    flex: {
-      flex: 1,
-    },
-  });
-
-// eslint-disable-next-line react/prop-types
-const Skeleton = ({ width, noStyle }) => {
-  const { colors } = useTheme();
-  const styles = createStyles(colors);
-
-  return (
-    <View style={[!noStyle && styles.valuesContainer]}>
-      <SkeletonPlaceholder>
-        <SkeletonPlaceholder.Item width={width} height={10} borderRadius={4} />
-      </SkeletonPlaceholder>
-    </View>
-  );
-};
-
-const TransactionReviewEIP1559 = ({
-  totalNative,
-  totalConversion,
-  totalMaxNative,
-  gasFeeNative,
-  gasFeeConversion,
-  gasFeeMaxNative,
-  gasFeeMaxConversion,
-  timeEstimate,
-  timeEstimateColor,
-  timeEstimateId,
+const TransactionReviewCard = ({
   primaryCurrency,
   chainId,
   onEdit,
@@ -103,27 +31,74 @@ const TransactionReviewEIP1559 = ({
   isAnimating,
   gasEstimationReady,
   legacy,
-}) => {
+  gasSelected,
+  gasObject,
+  onlyGas,
+  updateTransactionState,
+  multiLayerL1FeeTotal,
+  dappSuggestedGasPrice,
+  dappSuggestedEIP1559Gas,
+  totalMaxNative,
+  totalConversion,
+  totalNative,
+}: // gasFeeNative,
+TransactionReviewCardProps) => {
   const [showLearnMoreModal, setShowLearnMoreModal] = useState(false);
   const [
     isVisibleTimeEstimateInfoModal,
-    ,
     showTimeEstimateInfoModal,
     hideTimeEstimateInfoModal,
   ] = useModalHandler(false);
   const [isVisibleLegacyLearnMore, , showLegacyLearnMore, hideLegacyLearnMore] =
     useModalHandler(false);
   const toggleLearnMoreModal = useCallback(() => {
-    setShowLearnMoreModal((showLearnMoreModal) => !showLearnMoreModal);
-  }, []);
-  const { colors } = useTheme();
+    setShowLearnMoreModal(!showLearnMoreModal);
+  }, [showLearnMoreModal]);
+  const { colors } = useAppThemeFromContext() || mockTheme;
   const styles = createStyles(colors);
 
+  const gasTransaction = useGasTransaction({
+    onlyGas: !!onlyGas,
+    gasSelected,
+    legacy: !!legacy,
+    gasObject,
+    multiLayerL1FeeTotal,
+    dappSuggestedGasPrice,
+    dappSuggestedEIP1559Gas,
+  });
+
+  const {
+    renderableGasFeeMinNative,
+    renderableGasFeeMinConversion,
+    renderableGasFeeMaxNative,
+    renderableTotalMinNative,
+    renderableTotalMinConversion,
+    renderableTotalMaxNative,
+    renderableGasFeeMaxConversion,
+    timeEstimateColor,
+    timeEstimate,
+    timeEstimateId,
+    transactionFee,
+    transactionFeeFiat,
+    transactionTotalAmount,
+    transactionTotalAmountFiat,
+  } = gasTransaction;
+
+  // console.log('gasTransaction', gasTransaction);
+
+  totalMaxNative = renderableTotalMaxNative || totalMaxNative;
+  totalConversion = renderableTotalMinConversion || totalConversion;
+  totalNative = renderableTotalMinNative || totalNative;
+  // gasFeeNative = transactionTotalAmount || gasFeeNative;
+
+  useEffect(() => {
+    if (animateOnChange) {
+      updateTransactionState(gasTransaction);
+    }
+  }, [animateOnChange, gasTransaction, updateTransactionState]);
+
   const openLinkAboutGas = useCallback(
-    () =>
-      Linking.openURL(
-        'https://community.metamask.io/t/what-is-gas-why-do-transactions-take-so-long/3172',
-      ),
+    () => Linking.openURL(AppConstants.URLS.WHY_TRANSACTION_TAKE_TIME),
     [],
   );
 
@@ -133,30 +108,16 @@ const TransactionReviewEIP1559 = ({
 
   const isMainnet = isMainnetByChainId(chainId);
   const nativeCurrencySelected = primaryCurrency === 'ETH' || !isMainnet;
-  let gasFeePrimary,
-    gasFeeSecondary,
-    gasFeeMaxPrimary,
-    totalPrimary,
-    totalSecondary,
-    totalMaxPrimary;
-  if (nativeCurrencySelected) {
-    gasFeePrimary = gasFeeNative;
-    gasFeeSecondary = gasFeeConversion;
-    gasFeeMaxPrimary = gasFeeMaxNative;
-    totalPrimary = totalNative;
-    totalSecondary = totalConversion;
-    totalMaxPrimary = totalMaxNative;
-  } else {
-    gasFeePrimary = gasFeeConversion;
-    gasFeeSecondary = gasFeeNative;
-    gasFeeMaxPrimary = gasFeeMaxConversion;
-    totalPrimary = totalConversion;
-    totalSecondary = totalNative;
-    totalMaxPrimary = gasFeeMaxConversion;
-  }
 
-  const valueToWatchAnimation = `${gasFeeNative}${gasFeeMaxNative}`;
-  const isTestNetwork = isTestNet(chainId);
+  const switchNativeCurrencyDisplayOptions = (
+    nativeValue: string,
+    fiatValue: string,
+  ) => {
+    if (nativeCurrencySelected) return nativeValue;
+    return fiatValue;
+  };
+
+  const valueToWatchAnimation = `${renderableGasFeeMinNative}${renderableGasFeeMaxNative}`;
 
   return (
     <Summary style={styles.overview(noMargin)}>
@@ -214,7 +175,15 @@ const TransactionReviewEIP1559 = ({
                     adjustsFontSizeToFit
                     numberOfLines={2}
                   >
-                    {gasFeeSecondary}
+                    {legacy
+                      ? switchNativeCurrencyDisplayOptions(
+                          transactionFeeFiat,
+                          transactionFee,
+                        )
+                      : switchNativeCurrencyDisplayOptions(
+                          renderableGasFeeMinConversion,
+                          renderableGasFeeMinNative,
+                        )}
                   </Text>
                 </TouchableOpacity>
               )}
@@ -227,7 +196,7 @@ const TransactionReviewEIP1559 = ({
                 <Text
                   primary
                   bold
-                  upper={!isTestNetwork}
+                  upper
                   grey={!nativeCurrencySelected}
                   link={nativeCurrencySelected}
                   underline={nativeCurrencySelected}
@@ -236,12 +205,20 @@ const TransactionReviewEIP1559 = ({
                   adjustsFontSizeToFit
                   numberOfLines={2}
                 >
-                  {gasFeePrimary}
+                  {legacy
+                    ? switchNativeCurrencyDisplayOptions(
+                        transactionFee,
+                        transactionFeeFiat,
+                      )
+                    : switchNativeCurrencyDisplayOptions(
+                        renderableGasFeeMinNative,
+                        renderableGasFeeMinConversion,
+                      )}
                 </Text>
               </TouchableOpacity>
             </FadeAnimationView>
           ) : (
-            <Skeleton width={80} />
+            <SkeletonComponent width={80} />
           )}
         </View>
       </Summary.Row>
@@ -278,7 +255,7 @@ const TransactionReviewEIP1559 = ({
                 </View>
               </FadeAnimationView>
             ) : (
-              <Skeleton width={120} noStyle />
+              <SkeletonComponent width={120} noStyle />
             )}
             {gasEstimationReady ? (
               <FadeAnimationView
@@ -291,12 +268,15 @@ const TransactionReviewEIP1559 = ({
                     {strings('transaction_review_eip1559.max_fee')}:{' '}
                   </Text>
                   <Text small noMargin>
-                    {gasFeeMaxPrimary}
+                    {switchNativeCurrencyDisplayOptions(
+                      renderableGasFeeMaxNative,
+                      renderableGasFeeMaxConversion,
+                    )}
                   </Text>
                 </Text>
               </FadeAnimationView>
             ) : (
-              <Skeleton width={120} />
+              <SkeletonComponent width={120} />
             )}
           </View>
         </Summary.Row>
@@ -315,35 +295,54 @@ const TransactionReviewEIP1559 = ({
                   valueToWatch={valueToWatchAnimation}
                   animateOnChange={animateOnChange}
                 >
-                  {isMainnet && totalSecondary !== 'undefined' && (
-                    <Text
-                      grey
-                      upper
-                      right
-                      noMargin
-                      style={styles.amountContainer}
-                      adjustsFontSizeToFit
-                      numberOfLines={2}
-                    >
-                      {totalSecondary}
-                    </Text>
-                  )}
-
+                  {isMainnet &&
+                    switchNativeCurrencyDisplayOptions(
+                      renderableTotalMinConversion,
+                      renderableTotalMinNative,
+                    ) !== 'undefined' && (
+                      <Text
+                        grey
+                        upper
+                        right
+                        noMargin
+                        style={styles.amountContainer}
+                        adjustsFontSizeToFit
+                        numberOfLines={2}
+                      >
+                        {legacy
+                          ? switchNativeCurrencyDisplayOptions(
+                              transactionTotalAmountFiat,
+                              transactionTotalAmount,
+                            )
+                          : switchNativeCurrencyDisplayOptions(
+                              renderableTotalMinConversion,
+                              renderableTotalMinNative,
+                            )}
+                      </Text>
+                    )}
                   <Text
                     bold
                     primary
-                    upper={!isTestNetwork}
+                    upper
                     right
                     noMargin
                     style={[Device.isSmallDevice() && styles.flex]}
                     adjustsFontSizeToFit
                     numberOfLines={2}
                   >
-                    {totalPrimary}
+                    {legacy
+                      ? switchNativeCurrencyDisplayOptions(
+                          transactionTotalAmount,
+                          transactionTotalAmountFiat,
+                        )
+                      : switchNativeCurrencyDisplayOptions(
+                          totalNative,
+                          totalConversion,
+                        )}
                   </Text>
                 </FadeAnimationView>
               ) : (
-                <Skeleton width={80} />
+                <SkeletonComponent width={80} />
               )}
             </Summary.Row>
           </View>
@@ -360,12 +359,15 @@ const TransactionReviewEIP1559 = ({
                       {strings('transaction_review_eip1559.max_amount')}:
                     </Text>{' '}
                     <Text small noMargin>
-                      {totalMaxPrimary}
+                      {switchNativeCurrencyDisplayOptions(
+                        totalMaxNative,
+                        renderableGasFeeMaxConversion,
+                      )}
                     </Text>
                   </Text>
                 </FadeAnimationView>
               ) : (
-                <Skeleton width={120} />
+                <SkeletonComponent width={120} />
               )}
             </Summary.Row>
           )}
@@ -427,103 +429,4 @@ const TransactionReviewEIP1559 = ({
   );
 };
 
-TransactionReviewEIP1559.propTypes = {
-  /**
-   * Total value in native currency
-   */
-  totalNative: PropTypes.string,
-  /**
-   * Total value converted to chosen currency
-   */
-  totalConversion: PropTypes.string,
-  /**
-   * Total max value (amount + max fee) native
-   */
-  totalMaxNative: PropTypes.string,
-  /**
-   * Gas fee in native currency
-   */
-  gasFeeNative: PropTypes.string,
-  /**
-   * Gas fee converted to chosen currency
-   */
-  gasFeeConversion: PropTypes.string,
-  /**
-   * Maximum gas fee in native currency
-   */
-  gasFeeMaxNative: PropTypes.string,
-  /**
-   * Maximum gas fee onverted to chosen currency
-   */
-  gasFeeMaxConversion: PropTypes.string,
-  /**
-   * Selected primary currency
-   */
-  primaryCurrency: PropTypes.string,
-  /**
-   * A string representing the network chainId
-   */
-  chainId: PropTypes.string,
-  /**
-   * Function called when user clicks to edit the gas fee
-   */
-  onEdit: PropTypes.func,
-  /**
-   * String that represents the time estimates
-   */
-  timeEstimate: PropTypes.string,
-  /**
-   * String that represents the color of the time estimate
-   */
-  timeEstimateColor: PropTypes.string,
-  /**
-   * Time estimate name (unknown, low, medium, high, less_than, range)
-   */
-  timeEstimateId: PropTypes.string,
-  /**
-   * Boolean to determine if the total section should be hidden
-   */
-  hideTotal: PropTypes.bool,
-  /**
-   * Boolean to determine the container should have no margin
-   */
-  noMargin: PropTypes.bool,
-  /**
-   * Origin (hostname) of the dapp that suggested the gas fee
-   */
-  origin: PropTypes.string,
-  /**
-   * Function to call when update animation starts
-   */
-  onUpdatingValuesStart: PropTypes.func,
-  /**
-   * Function to call when update animation ends
-   */
-  onUpdatingValuesEnd: PropTypes.func,
-  /**
-   * If the values should animate upon update or not
-   */
-  animateOnChange: PropTypes.bool,
-  /**
-   * Boolean to determine if the animation is happening
-   */
-  isAnimating: PropTypes.bool,
-  /**
-   * If loading should stop
-   */
-  gasEstimationReady: PropTypes.bool,
-  /**
-   * If should show legacy gas
-   */
-  legacy: PropTypes.bool,
-  /**
-   * If it's a eip1559 network and dapp suggest legact gas then it should show a warning
-   */
-  originWarning: PropTypes.bool,
-};
-
-const mapStateToProps = (state) => ({
-  chainId: state.engine.backgroundState.NetworkController.provider.chainId,
-});
-
-export default connect(mapStateToProps)(TransactionReviewEIP1559);
+export default TransactionReviewCard;
